refactor: 🔄 type tests: src => type
